Move tx, rx out of State, restructure main loop.
Having tx, rx in State caused issues with the borrow checker, and we don't really need these
channels in the State struct.  Moved them back out.

Rewrote the if statement in the main loop where we check the output mode to see if we should be doing
anything with this output.  Using a match statement looks cleaner.