Add 2nd pair of channels to enable bidirectional communication.
This allows the main thread to send success/failure messages back to the dbus thread after executing whatever command
was sent from the dbus thread.

We may want to have a timeout for the recv call in the dbus thread in case the main thread crashes before it can send
a response back to the dbus thread.